fix aseprite command on linux

+++ addons/nklbdev.aseprite_importers/aseprite_image_format_loader_extension.gd
@@ -7,15 +7,12 @@ const __recognized_extensions: PackedStringArray = ["ase", "aseprite"]
 func _get_recognized_extensions() -> PackedStringArray:
 	return __recognized_extensions
 
-func _load_image(image: Image, file_access: FileAccess, flags: ImageFormatLoader.LoaderFlags, scale: float) -> Error:
+func _load_image(image: Image, file_access: FileAccess, flags: int, scale: float) -> Error:
 	flags = flags as ImageFormatLoader.LoaderFlags
 
 	var source_file_path: String = ProjectSettings.globalize_path(file_access.get_path_absolute())
 	var global_png_path: String = source_file_path + ".png"
 	var aseprite_executable_path: String = ProjectSettings.get_setting(Common.ASEPRITE_EXECUTABLE_PATH_SETTING_NAME)
-	if not FileAccess.file_exists(aseprite_executable_path):
-		push_error("Cannot find Aseprite executable. Check Aseprite executable path in project settings.")
-		return ERR_BUG
 
 	var command_line_params: PackedStringArray = PackedStringArray([
 		"--batch",
@@ -29,6 +26,9 @@ func _load_image(image: Image, file_access: FileAccess, flags: ImageFormatLoader
 	var err: Error = OS.execute(
 		ProjectSettings.get_setting(Common.ASEPRITE_EXECUTABLE_PATH_SETTING_NAME),
 		command_line_params, output, true)
+	if err:
+		push_error("There was an error while executing aseprite command: %s" % error_string(err))
+		return err
 
 	image.load_png_from_buffer(FileAccess.get_file_as_bytes(global_png_path))
 	DirAccess.remove_absolute(global_png_path)
